+++ lib/pragma_design_system.dart
@@ -38,6 +38,7 @@ export 'src/widgets/pragma_icon_button_widget.dart';
 export 'src/widgets/pragma_input_widget.dart';
 export 'src/widgets/pragma_loading_widget.dart';
 export 'src/widgets/pragma_logo_widget.dart';
+export 'src/widgets/pragma_pagination_widget.dart';
 export 'src/widgets/pragma_radio_button_widget.dart';
 export 'src/widgets/pragma_search_widget.dart';
 export 'src/widgets/pragma_stepper_widget.dart';